Improves the image viewer demo by adding rollover, screen refresh controls, and invert function.
The viewer will now jump to image 0 if down is pressed on the last photo, or jump to the last if up is pressed on image 0.
To invert the display, press B.
To switch between TURBO and FAST refresh speeds, press C.
To jump to / switch between MEDIUM and SLOW speeds, press C while holding the USER button.