Update network_manager.py
In order to successfully connect to more complicated AP setups (ie WPA-WPA2-WPA3) mesh networks, the power save mode needs to be in place. It is possible that a different mode can be set once a socket is opened. For now this is necessary for a range of ASUS routers/APs to be able to negotiate a connection.